Fixes and improvements
