+++ src/context/Localization/language/es.json
@@ -7,6 +7,7 @@
     "done": "Done",
     "create": "Create",
     "cancel": "Cancel",
+    "Goback": "Go back",
     "clear": "Clear",
     "settings": "Settings",
     "add_new": "Add New",
